feat: deprecate `node` in favour of `chain-spec-generator`
