+++ chain-spec-generator/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+authors.workspace = true
+description = "Chain Spec Generator for Kreivo"
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "chain-spec-generator"
+repository.workspace = true
+version = "0.13.3"
+
+[dependencies]
+clap = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+serde = { workspace = true, features = ["derive"] }
+
+kreivo-runtime = { workspace = true, default-features = true }
+
+cumulus-primitives-core.workspace = true
+parachains-common.workspace = true
+sp-core.workspace = true
+sp-runtime.workspace = true
+sc-chain-spec.workspace = true
+sc-network.workspace = true
+xcm.workspace = true
+
+[features]
+runtime-benchmarks = [
+  "kreivo-runtime/runtime-benchmarks"
+]
+paseo = [
+  "kreivo-runtime/paseo"
+]
