--- node/Cargo.toml
@@ -1,99 +1,0 @@
-[package]
-authors.workspace = true
-build = "build.rs"
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-name = "virto-node"
-publish = false
-repository.workspace = true
-version = "0.13.3"
-
-[[bin]]
-name = "virto-node"
-path = "src/main.rs"
-
-[dependencies]
-clap = { version = "4.5.3", features = ["derive"] }
-codec = { package = "parity-scale-codec", version = "3.6.12" }
-docify = "0.2.8"
-futures = "0.3.28"
-jsonrpsee = { version = "0.22", features = ["server"] }
-log = { workspace = true, default-features = true }
-serde = { features = ["derive"], workspace = true, default-features = true }
-serde_json = { workspace = true, default-features = true }
-
-# Local
-kreivo-runtime = { workspace = true, default-features = true }
-parachains-common.workspace = true
-
-# Substrate
-frame-benchmarking.workspace = true
-frame-benchmarking-cli.workspace = true
-pallet-transaction-payment-rpc.workspace = true
-substrate-prometheus-endpoint.workspace = true
-sc-basic-authorship.workspace = true
-sc-chain-spec.workspace = true
-sc-cli.workspace = true
-sc-client-api.workspace = true
-sc-consensus.workspace = true
-sc-executor.workspace = true
-sc-network.workspace = true
-sc-network-sync.workspace = true
-sc-offchain.workspace = true
-sc-rpc.workspace = true
-sc-service.workspace = true
-sc-sysinfo.workspace = true
-sc-telemetry.workspace = true
-sc-tracing.workspace = true
-sc-transaction-pool.workspace = true
-sc-transaction-pool-api.workspace = true
-sp-api.workspace = true
-sp-block-builder.workspace = true
-sp-blockchain.workspace = true
-sp-consensus-aura.workspace = true
-sp-core.workspace = true
-sp-io.workspace = true
-sp-keystore.workspace = true
-sp-runtime.workspace = true
-sp-timestamp.workspace = true
-substrate-frame-rpc-system.workspace = true
-
-# Polkadot
-polkadot-cli.workspace = true
-polkadot-primitives.workspace = true
-xcm = { workspace = true, default-features = false }
-
-# Cumulus
-color-print = "0.3.4"
-cumulus-client-cli.workspace = true
-cumulus-client-collator.workspace = true
-cumulus-client-consensus-aura.workspace = true
-cumulus-client-consensus-common.workspace = true
-cumulus-client-consensus-proposer.workspace = true
-cumulus-client-service.workspace = true
-cumulus-primitives-core.workspace = true
-cumulus-primitives-parachain-inherent.workspace = true
-cumulus-relay-chain-interface.workspace = true
-
-[build-dependencies]
-substrate-build-script-utils.workspace = true
-
-[features]
-default = []
-paseo = ["kreivo-runtime/paseo"]
-runtime-benchmarks = [
-  "cumulus-primitives-core/runtime-benchmarks",
-  "frame-benchmarking-cli/runtime-benchmarks",
-  "frame-benchmarking/runtime-benchmarks",
-  "kreivo-runtime/runtime-benchmarks",
-  "polkadot-cli/runtime-benchmarks",
-  "polkadot-primitives/runtime-benchmarks",
-  "sc-service/runtime-benchmarks",
-  "sp-runtime/runtime-benchmarks",
-]
-try-runtime = [
-  "kreivo-runtime/try-runtime",
-  "polkadot-cli/try-runtime",
-  "sp-runtime/try-runtime",
-]
