--- node/src/main.rs
@@ -1,13 +1,0 @@
-//! Substrate Parachain Node Template CLI
-
-#![warn(missing_docs)]
-
-mod chain_spec;
-mod cli;
-mod command;
-mod rpc;
-mod service;
-
-fn main() -> sc_cli::Result<()> {
-	command::run()
-}
